requirements.txt
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.47.0
fsspec==2024.2.0
idna==3.6
imbalanced-learn==0.12.0
imblearn==0.0
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opt-einsum==3.3.0
owlready2==0.46
packaging==23.2
pandas==2.1.4
patsy==0.5.6
pgmpy==0.1.24
pillow==10.2.0
pip==24.0
prettytable==3.10.0
pyparsing==3.1.1
pyswip==0.2.10
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.2
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
spotipy==2.23.0
statsmodels==0.14.1
sympy==1.12
threadpoolctl==3.2.0
torch==2.2.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2023.4
urllib3==2.2.1
seaborn==0.13.2
scikit-optimize==0.10.1 
python-constraint==1.4.0
csp==1.16.0